+++ engine/server/engine/main.go
@@ -346,6 +346,8 @@ func getEnclaveManager(
 	switch kurtosisBackendType {
 	case args.KurtosisBackendType_Docker:
 		apiContainerKurtosisBackendConfigSupplier = api_container_launcher.NewDockerKurtosisBackendConfigSupplier()
+	case args.KurtosisBackendType_Podman:
+		apiContainerKurtosisBackendConfigSupplier = api_container_launcher.NewPodmanKurtosisBackendConfigSupplier()
 	case args.KurtosisBackendType_Kubernetes:
 		kurtosisLocalBackendConfigKubernetesType, ok := kurtosisLocalBackendConfig.(kurtosis_backend_config.KubernetesBackendConfig)
 		if !ok {
@@ -387,6 +389,11 @@ func getKurtosisBackend(ctx context.Context, kurtosisBackendType args.KurtosisBa
 		kurtosisBackend, err = backend_creator.GetDockerKurtosisBackend(apiContainerModeArgsForKurtosisBackend, remoteBackendConfigMaybe)
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred getting local Docker Kurtosis backend")
+		}
+	case args.KurtosisBackendType_Podman:
+		kurtosisBackend, err = backend_creator.GetPodmanKurtosisBackend(apiContainerModeArgsForKurtosisBackend, remoteBackendConfigMaybe)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting local Podman Kurtosis backend")
 		}
 	case args.KurtosisBackendType_Kubernetes:
 		if remoteBackendConfigMaybe != nil {